requirements.txt
streamlit
pandas
numpy
langchain
pypdf
ipykernel
langchain-community
nltk
langchain_google_genai
python-dotenv
chromadb
cohere